# memory: optimize peer store performance