Removing `defer`s whenever not too inconvenient:

```
benchmark                                                            old ns/op     new ns/op     delta
BenchmarkPeerStore_PutSeeder-8                                       894           890           -0.45%
BenchmarkPeerStore_PutSeeder1KInfohash-8                             1164          758           -34.88%
BenchmarkPeerStore_PutSeeder1KSeeders-8                              941           766           -18.60%
BenchmarkPeerStore_PutSeeder1KInfohash1KSeeders-8                    1160          946           -18.45%
BenchmarkPeerStore_PutDeleteSeeder-8                                 3479          3918          +12.62%
BenchmarkPeerStore_PutDeleteSeeder1KInfohash-8                       3315          3142          -5.22%
BenchmarkPeerStore_PutDeleteSeeder1KSeeders-8                        3696          3702          +0.16%
BenchmarkPeerStore_PutDeleteSeeder1KInfohash1KSeeders-8              3345          3436          +2.72%
BenchmarkPeerStore_DeleteSeederNonExist-8                            654           545           -16.67%
BenchmarkPeerStore_DeleteSeederNonExist1KInfohash-8                  664           538           -18.98%
BenchmarkPeerStore_DeleteSeederNonExist1KSeeders-8                   628           576           -8.28%
BenchmarkPeerStore_DeleteSeederNonExist1KInfohash1KSeeders-8         636           607           -4.56%
BenchmarkPeerStore_PutGraduateDeleteLeecher-8                        6212          4698          -24.37%
BenchmarkPeerStore_PutGraduateDeleteLeecher1KInfohash-8              5781          4950          -14.37%
BenchmarkPeerStore_PutGraduateDeleteLeecher1KSeeders-8               5982          5271          -11.89%
BenchmarkPeerStore_PutGraduateDeleteLeecher1KInfohash1KSeeders-8     6511          4403          -32.38%
BenchmarkPeerStore_GraduateLeecherNonExist-8                         1116          1022          -8.42%
BenchmarkPeerStore_GraduateLeecherNonExist1KInfohash-8               1556          1092          -29.82%
BenchmarkPeerStore_GraduateLeecherNonExist1KSeeders-8                1332          1141          -14.34%
BenchmarkPeerStore_GraduateLeecherNonExist1KInfohash1KSeeders-8      1653          1266          -23.41%
BenchmarkPeerStore_AnnouncePeers-8                                   15380         15897         +3.36%
BenchmarkPeerStore_AnnouncePeers1KInfohash-8                         19741         19979         +1.21%
BenchmarkPeerStore_AnnouncePeersSeeder-8                             15607         15536         -0.45%
BenchmarkPeerStore_AnnouncePeersSeeder1KInfohash-8                   19844         20792         +4.78%
BenchmarkPeerStore_GetSeeders-8                                      127890        127051        -0.66%
BenchmarkPeerStore_GetSeeders1KInfohash-8                            149643        148978        -0.44%
BenchmarkPeerStore_NumSeeders-8                                      236           120           -49.15%
BenchmarkPeerStore_NumSeeders1KInfohash-8                            274           157           -42.70%
```