@jzelinskie 2016-08-01
The code looks fine, I'll merge it if you change the commit message to be a bit more descriptive.

For example:
```
memory: removed usage of defer key

Removing the overhead of the defer keyword improved performance of some
benchmarks up to 35%
```